Self review refactor and tidy
